requirements.txt
attrs==21.4.0
black==22.6.0
boto3==1.24.21
botocore==1.27.21
Cerberus==1.3.4
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
coverage==6.4.1
cryptography==37.0.2
Flask==2.1.2
idna==3.3
importlib-metadata==4.12.0
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jwt==1.3.1
MarkupSafe==2.1.1
moto==3.1.16
mypy-extensions==0.4.3
numpy==1.23.0
packaging==21.3
pandas==1.4.3
pathspec==0.9.0
platformdirs==2.5.2
pluggy==1.0.0
py==1.11.0
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
pytest==7.1.2
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.1
responses==0.21.0
s3transfer==0.6.0
six==1.16.0
tomli==2.0.1
typing-extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.2
xmltodict==0.13.0
zipp==3.8.0
